+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Add `Regions.get_item`
 * GDAL COG-backend for CocoFrames
 * `CocoDataset.remove_images`
+* `CocoDataset._alias_to_cat` looks up categories via the alias foreign key
 
 ### Changed
 * Removed redundant items from `CategoryTree` serialization in `__json__` and `__getstate__`.